fix: finish app refactor
