+++ src/main/java/edu/washu/bank/App.java
@@ -1,14 +1,10 @@
 package edu.washu.bank;
 
+import edu.washu.bank.cli.BankCli;
 import edu.washu.bank.core.Bank;
-import edu.washu.bank.model.Account;
-import edu.washu.bank.model.AccountType;
-import edu.washu.bank.model.Transaction;
 import edu.washu.bank.persistence.SqliteBankStore;
 import edu.washu.bank.service.AccountService;
 
-import java.math.BigDecimal;
-import java.util.List;
 import java.nio.file.Path;
 import java.sql.SQLException;
 
@@ -26,401 +22,6 @@ public class App {
         }
 
         AccountService accountService = new AccountService(bank);
-
-        if (args.length == 0 || "help".equalsIgnoreCase(args[0])) {
-            printUsage(dbPath);
-            return;
-        }
-
-        if ("create-account".equalsIgnoreCase(args[0])) {
-            runCreateAccount(store, bank, accountService, args);
-            return;
-        }
-
-        if ("check-balance".equalsIgnoreCase(args[0])) {
-            runCheckBalance(accountService, args);
-            return;
-        }
-
-        if ("transaction-history".equalsIgnoreCase(args[0])) {
-            runTransactionHistory(accountService, args);
-            return;
-        }
-
-        if ("deposit".equalsIgnoreCase(args[0])) {
-            runDeposit(store, bank, accountService, args);
-            return;
-        }
-
-        if ("withdraw".equalsIgnoreCase(args[0])) {
-            runWithdraw(store, bank, accountService, args);
-            return;
-        }
-
-        if ("close-account".equalsIgnoreCase(args[0])) {
-            runCloseAccount(store, bank, accountService, args);
-            return;
-        }
-
-        if ("transfer".equalsIgnoreCase(args[0])) {
-            runTransfer(store, bank, accountService, args);
-            return;
-        }
-
-        if ("collect-fee".equalsIgnoreCase(args[0])) {
-            runCollectFee(store, bank, accountService, args);
-            return;
-        }
-
-        if ("add-interest".equalsIgnoreCase(args[0])) {
-            runAddInterest(store, bank, accountService, args);
-            return;
-        }
-
-        if ("clear-data".equalsIgnoreCase(args[0])) {
-            runClearData(store, dbPath);
-            return;
-        }
-
-        System.out.println("Unknown command: " + args[0]);
-        printUsage(dbPath);
-    }
-
-    private static void runCreateAccount(
-            SqliteBankStore store,
-            Bank bank,
-            AccountService accountService,
-            String[] args
-    ) {
-        if (args.length != 4) {
-            System.out.println("Invalid arguments for create-account.");
-            printUsage(SqliteBankStore.resolveDatabasePath());
-            return;
-        }
-
-        String customerId = args[1];
-
-        AccountType accountType;
-        try {
-            accountType = AccountType.valueOf(args[2].toUpperCase());
-        } catch (IllegalArgumentException ex) {
-            System.out.println("Invalid account type: " + args[2] + ". Use CHECKING or SAVINGS.");
-            return;
-        }
-
-        BigDecimal openingDeposit;
-        try {
-            openingDeposit = new BigDecimal(args[3]);
-        } catch (NumberFormatException ex) {
-            System.out.println("Invalid opening deposit: " + args[3]);
-            return;
-        }
-
-        try {
-            Account account = accountService.createAdditionalAccount(customerId, accountType, openingDeposit);
-            store.saveFullState(bank);
-            System.out.println(
-                    "Created account " + account.getId()
-                            + " (" + account.getType() + ") for customer " + account.getCustomerId()
-                            + " with opening balance " + account.getBalance()
-            );
-        } catch (RuntimeException ex) {
-            System.out.println(ex.getMessage());
-        } catch (SQLException ex) {
-            System.err.println("Database error: " + ex.getMessage());
-        }
-    }
-
-    private static void runCheckBalance(AccountService accountService, String[] args) {
-        if (args.length != 2) {
-            System.out.println("Invalid arguments for check-balance.");
-            printUsage(SqliteBankStore.resolveDatabasePath());
-            return;
-        }
-
-        String accountId = args[1];
-
-        try {
-            BigDecimal balance = accountService.getBalance(accountId);
-            System.out.println("Account " + accountId + " balance: " + balance);
-        } catch (RuntimeException ex) {
-            System.out.println(ex.getMessage());
-        }
-    }
-
-    private static void runTransactionHistory(AccountService accountService, String[] args) {
-        if (args.length != 2) {
-            System.out.println("Invalid arguments for transaction-history.");
-            printUsage(SqliteBankStore.resolveDatabasePath());
-            return;
-        }
-
-        String accountId = args[1];
-        try {
-            List<Transaction> history = accountService.getTransactionHistory(accountId);
-            System.out.println("Transaction history for account " + accountId + ":");
-            if (history.isEmpty()) {
-                System.out.println("  No transactions found.");
-                return;
-            }
-            for (Transaction transaction : history) {
-                String related = transaction.getRelatedAccountId() == null
-                        ? ""
-                        : " | related account: " + transaction.getRelatedAccountId();
-                System.out.println(
-                        "  " + transaction.getId()
-                                + " | " + transaction.getType()
-                                + " | amount: " + transaction.getAmount()
-                                + " | balance after: " + transaction.getBalanceAfter()
-                                + related
-                                + " | " + transaction.getDescription()
-                );
-            }
-        } catch (RuntimeException ex) {
-            System.out.println(ex.getMessage());
-        }
-    }
-
-    private static void runDeposit(
-            SqliteBankStore store,
-            Bank bank,
-            AccountService accountService,
-            String[] args
-    ) {
-        if (args.length != 3) {
-            System.out.println("Invalid arguments for deposit.");
-            printUsage(SqliteBankStore.resolveDatabasePath());
-            return;
-        }
-
-        String accountId = args[1];
-
-        BigDecimal amount;
-        try {
-            amount = new BigDecimal(args[2]);
-        } catch (NumberFormatException ex) {
-            System.out.println("Invalid deposit amount: " + args[2]);
-            return;
-        }
-
-        try {
-            Account updatedAccount = accountService.depositIntoExistingAccount(accountId, amount);
-            store.saveFullState(bank);
-            System.out.println(
-                    "Deposited " + amount
-                            + " into account " + updatedAccount.getId()
-                            + ". New balance: " + updatedAccount.getBalance()
-            );
-        } catch (RuntimeException ex) {
-            System.out.println(ex.getMessage());
-        } catch (SQLException ex) {
-            System.err.println("Database error: " + ex.getMessage());
-        }
-    }
-
-    private static void runWithdraw(
-            SqliteBankStore store,
-            Bank bank,
-            AccountService accountService,
-            String[] args
-    ) {
-        if (args.length != 3) {
-            System.out.println("Invalid arguments for withdraw.");
-            printUsage(SqliteBankStore.resolveDatabasePath());
-            return;
-        }
-
-        String accountId = args[1];
-
-        BigDecimal amount;
-        try {
-            amount = new BigDecimal(args[2]);
-        } catch (NumberFormatException ex) {
-            System.out.println("Invalid withdraw amount: " + args[2]);
-            return;
-        }
-
-        try {
-            Account updatedAccount = accountService.withdraw(accountId, amount);
-            store.saveFullState(bank);
-            System.out.println(
-                    "Withdrew " + amount
-                            + " from account " + updatedAccount.getId()
-                            + ". New balance: " + updatedAccount.getBalance()
-            );
-        } catch (RuntimeException ex) {
-            System.out.println(ex.getMessage());
-        } catch (SQLException ex) {
-            System.err.println("Database error: " + ex.getMessage());
-        }
-    }
-
-    private static void runCloseAccount(
-            SqliteBankStore store,
-            Bank bank,
-            AccountService accountService,
-            String[] args
-    ) {
-        if (args.length != 2) {
-            System.out.println("Invalid arguments for close-account.");
-            printUsage(SqliteBankStore.resolveDatabasePath());
-            return;
-        }
-
-        String accountId = args[1];
-        try {
-            BigDecimal cashOutAmount = accountService.closeAccount(accountId);
-            store.saveFullState(bank);
-            System.out.println(
-                    "Closed account " + accountId + ". Cash-out amount: " + cashOutAmount
-            );
-        } catch (RuntimeException ex) {
-            System.out.println(ex.getMessage());
-        } catch (SQLException ex) {
-            System.err.println("Database error: " + ex.getMessage());
-        }
-    }
-
-    private static void runTransfer(
-            SqliteBankStore store,
-            Bank bank,
-            AccountService accountService,
-            String[] args
-    ) {
-        if (args.length != 4) {
-            System.out.println("Invalid arguments for transfer.");
-            printUsage(SqliteBankStore.resolveDatabasePath());
-            return;
-        }
-
-        BigDecimal amount;
-        try {
-            amount = new BigDecimal(args[3]);
-        } catch (NumberFormatException ex) {
-            System.out.println("Invalid transfer amount: " + args[3]);
-            return;
-        }
-
-        try {
-            accountService.transfer(args[1], args[2], amount);
-            store.saveFullState(bank);
-            System.out.println(
-                    "Transferred " + amount + " from " + args[1] + " to " + args[2] + "."
-            );
-        } catch (RuntimeException ex) {
-            System.out.println(ex.getMessage());
-        } catch (SQLException ex) {
-            System.err.println("Database error: " + ex.getMessage());
-        }
-    }
-
-    private static void runCollectFee(
-            SqliteBankStore store,
-            Bank bank,
-            AccountService accountService,
-            String[] args
-    ) {
-        if (args.length != 5) {
-            System.out.println("Invalid arguments for collect-fee.");
-            printUsage(SqliteBankStore.resolveDatabasePath());
-            return;
-        }
-
-        BigDecimal amount;
-        try {
-            amount = new BigDecimal(args[4]);
-        } catch (NumberFormatException ex) {
-            System.out.println("Invalid fee amount: " + args[4]);
-            return;
-        }
-
-        try {
-            Account updatedAccount = accountService.collectFee(args[1], args[2], args[3], amount);
-            store.saveFullState(bank);
-            System.out.println(
-                    "Collected fee of " + amount + " from account " + updatedAccount.getId()
-                            + ". New balance: " + updatedAccount.getBalance()
-            );
-        } catch (RuntimeException ex) {
-            System.out.println(ex.getMessage());
-        } catch (SQLException ex) {
-            System.err.println("Database error: " + ex.getMessage());
-        }
-    }
-
-    private static void runAddInterest(
-            SqliteBankStore store,
-            Bank bank,
-            AccountService accountService,
-            String[] args
-    ) {
-        if (args.length != 5) {
-            System.out.println("Invalid arguments for add-interest.");
-            printUsage(SqliteBankStore.resolveDatabasePath());
-            return;
-        }
-
-        BigDecimal amount;
-        try {
-            amount = new BigDecimal(args[4]);
-        } catch (NumberFormatException ex) {
-            System.out.println("Invalid interest amount: " + args[4]);
-            return;
-        }
-
-        try {
-            Account updatedAccount = accountService.addInterest(args[1], args[2], args[3], amount);
-            store.saveFullState(bank);
-            System.out.println(
-                    "Added interest of " + amount + " to account " + updatedAccount.getId()
-                            + ". New balance: " + updatedAccount.getBalance()
-            );
-        } catch (RuntimeException ex) {
-            System.out.println(ex.getMessage());
-        } catch (SQLException ex) {
-            System.err.println("Database error: " + ex.getMessage());
-        }
-    }
-
-    private static void runClearData(SqliteBankStore store, Path dbPath) {
-        try {
-            store.clearAllAndReseed();
-            System.out.println("Cleared database and re-seeded demo customer CUST-001.");
-            System.out.println("Database file: " + dbPath.toAbsolutePath().normalize());
-        } catch (SQLException ex) {
-            System.err.println("Database error: " + ex.getMessage());
-        }
-    }
-
-    private static void printUsage(Path dbPath) {
-        System.out.println("Bank CLI (SQLite: " + dbPath.toAbsolutePath().normalize() + ")");
-        System.out.println("Override path: -Dbank.db.file=/path/to/bank.db");
-        System.out.println("Seeded customer for demo: CUST-001");
-        System.out.println(
-                "Seeded admin for demo: "
-                        + SqliteBankStore.SEEDED_ADMIN_USERNAME
-                        + " / "
-                        + SqliteBankStore.SEEDED_ADMIN_PASSWORD
-        );
-        System.out.println("Usage:");
-        System.out.println("  create-account <customerId> <CHECKING|SAVINGS> <openingDeposit>");
-        System.out.println("  transaction-history <accountId>");
-        System.out.println("  deposit <accountId> <amount>");
-        System.out.println("  withdraw <accountId> <amount>");
-        System.out.println("  close-account <accountId>");
-        System.out.println("  transfer <fromAccountId> <toAccountId> <amount>");
-        System.out.println("  collect-fee <adminUsername> <adminPassword> <accountId> <amount>");
-        System.out.println("  add-interest <adminUsername> <adminPassword> <accountId> <amount>");
-        System.out.println("  check-balance <accountId>");
-        System.out.println("  clear-data");
-        System.out.println("Examples:");
-        System.out.println("  create-account CUST-001 CHECKING 100.00");
-        System.out.println("  check-balance ACC-0001");
-        System.out.println("  deposit ACC-0001 50.00");
-        System.out.println("  withdraw ACC-0001 25.00");
-        System.out.println("  transaction-history ACC-0001");
-        System.out.println("  transfer ACC-0001 ACC-0002 10.00");
-        System.out.println("  collect-fee admin admin123 ACC-0001 5.00");
-        System.out.println("  add-interest admin admin123 ACC-0001 3.00");
+        new BankCli(store, bank, accountService, dbPath).run(args);
     }
 }
